Add user authetication

+++ routes/api.php
@@ -21,23 +21,27 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
+Route::post('/register', [UserController::class, 'createUser']);
+Route::post('/login', [UserController::class, 'login']);
+Route::post('/logout', [UserController::class, 'logout']);
+
 /* Return data */
-Route::get('/users', [UserController::class, 'index']);
+//Route::get('/users', [UserController::class, 'index']);
 Route::get('/products', [ProductController::class, 'index']);
 Route::get('/products/{id}', [ProductController::class, 'getProductById']);
-Route::get('/users/{name}', [UserController::class, 'getUserByName']);
+//Route::get('/users/{name}', [UserController::class, 'getUserByName']);
 
 /* Create a new user */
-Route::post('/users', [UserController::class, 'createUser']);
+//Route::post('/users', [UserController::class, 'createUser']);
 /* Create a new product */
-Route::post('/products', [ProductController::class, 'createProduct']);
+Route::middleware('auth:sanctum')->post('/products', [ProductController::class, 'createProduct']);
 
 /* Update a user */
 Route::put('/users/{name}', [UserController::class, 'updateUser']);
 /* Update a user */
-Route::put('/products/{id}', [ProductController::class, 'updateProduct']);
+Route::middleware('auth:sanctum')->put('/products/{id}', [ProductController::class, 'updateProduct']);
 
 /* Delete a user */
 Route::delete('/users/{name}', [UserController::class, 'deleteUser']);
 /* Delete a product */
-Route::delete('/products/{id}', [ProductController::class, 'deleteProduct']);
+Route::middleware('auth:sanctum')->delete('/products/{id}', [ProductController::class, 'deleteProduct']);
